// src/generated/meta.ts
export type ConfigKey
  = | 'framework.enabled'
    | 'react.languageScopes'
    | 'react.importReactOnTop'
    | 'vue.languageScopes'
    | 'vue.uniappCodeSnippets'
    | 'vue.vuexCodeSnippets'
    | 'vue.useTemplateSnippets';

export interface ConfigKeyTypeMap {
  'framework.enabled': ('vue' | 'react' | 'nest' | 'vitest')[];
  'react.languageScopes': string;
  'react.importReactOnTop': boolean;
  'vue.languageScopes': string;
  'vue.uniappCodeSnippets': boolean;
  'vue.vuexCodeSnippets': boolean;
  'vue.useTemplateSnippets': string[];
}
// ...
export interface ConfigItem<T extends keyof ConfigKeyTypeMap> {
  key: T;
  default: ConfigKeyTypeMap[T];
}

/**
 * Configs map registered by `sankeyangshu.vscode-javascript-snippets`
 */
export const configs = {
  /**
   * Select which framework snippets to enable
   * @key `framework.enabled`
   * @default `["vue","react","nest","vitest"]`
   * @type `array`
   */
  frameworkEnabled: {
    key: 'framework.enabled',
    default: ['vue', 'react', 'nest', 'vitest'],
  } as ConfigItem<'framework.enabled'>,
  /**
   *
   * @key `react.languageScopes`
   * @default `"typescript,typescriptreact,javascript,javascriptreact"`
   * @type `string`
   */
  reactLanguageScopes: {
    key: 'react.languageScopes',
    default: 'typescript,typescriptreact,javascript,javascriptreact',
  } as ConfigItem<'react.languageScopes'>,
  /**
   *
   * @key `react.importReactOnTop`
   * @default `false`
   * @type `boolean`
   */
  reactImportReactOnTop: {
    key: 'react.importReactOnTop',
    default: false,
  } as ConfigItem<'react.importReactOnTop'>,
  /**
   *
   * @key `vue.languageScopes`
   * @default `"vue,typescript,javascript"`
   * @type `string`
   */
  vueLanguageScopes: {
    key: 'vue.languageScopes',
    default: 'vue,typescript,javascript',
  } as ConfigItem<'vue.languageScopes'>,
  /**
   *
   * @key `vue.uniappCodeSnippets`
   * @default `true`
   * @type `boolean`
   */
  vueUniappCodeSnippets: {
    key: 'vue.uniappCodeSnippets',
    default: true,
  } as ConfigItem<'vue.uniappCodeSnippets'>,
  /**
   *
   * @key `vue.vuexCodeSnippets`
   * @default `false`
   * @type `boolean`
   */
  vueVuexCodeSnippets: {
    key: 'vue.vuexCodeSnippets',
    default: false,
  } as ConfigItem<'vue.vuexCodeSnippets'>,
  /**
   * Defines the template snippets for which the snippets will be available
   * @key `vue.useTemplateSnippets`
   * @default `["vinit","vinit-scss","vbase","vbase-scss","vts","vts-scss"]`
   * @type `array`
   */
  vueUseTemplateSnippets: {
    key: 'vue.useTemplateSnippets',
    default: ['vinit', 'vinit-scss', 'vbase', 'vbase-scss', 'vts', 'vts-scss'],
  } as ConfigItem<'vue.useTemplateSnippets'>,
};